# Add new `Performance/ReverseFirst` cop